Apriori Algorithm
Apriori is a data mining technique designed to operate on databases containing transactions (for example, collections of items bought by customers, or details of a website frequentation).It proceeds by identifying the frequent individual items in the database and extending them to larger and larger item sets as long as those item sets appear sufficiently often in the database. In other words, a itemset can be considered for candidature of frequent set only if its all subsets are also the frequent itemsets. Apriori algorithm has applications in domains such as market basket analysis.
Further Apriori algorithm can be exploited to find (hidden) association rules, from a itemset -> item, in which the item is present in every transaction that contains the corresponding itemsets.
Note that an item is frequent or not is decided by the value of support parameter which is evaluated by number of occurrences/total number of transactions or sometimes just the number of transactions that contain the item.